# src/main/resources/static/api.yaml
openapi: 3.0.1
info:
  title: Requisites API
  version: '1.0'
servers:
  - url: http://localhost:8086
    description: Dev server
paths:
  /requisites:
    get:
      summary: Получение всех requisites
      operationId: getRequisites
      tags:
        - requisites
      responses:
        '200':
          description: Список реквизитов
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/Requisites"
  /requisites/{externalId}:
    get:
      summary: Получение requisites по externalId
      operationId: getRequisitesById
      tags:
        - requisites
      parameters:
        - name: externalId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Реквизиты
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Requisites"
  /client:
    get:
      operationId: getClient
      tags:
        - client-req-controller
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/Client"
  /client/{id}:
    get:
      operationId: getClientById
      tags:
        - client-req-controller
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Client"
components:
  schemas:
    Requisites:
      type: object
      properties:
        id:
          type: string
          format: uuid
        externalId:
          type: string
        clientId:
          type: string
        nameCompany:
          type: string
        inn:
          type: integer
          format: int64
        kpp:
          type: integer
          format: int32
        ogrn:
          type: integer
          format: int64
        businessAddress:
          type: string
        address:
          type: string
        rcbic:
          type: integer
          format: int32
        corrAss:
          type: number
          format: double
        ass:
          type: number
          format: double
        bankName:
          type: string
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time
    Client:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        requisitesList:
          type: array
          items:
            "$ref": "#/components/schemas/Requisites"